# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This contains the commn REST api to support a modern cricket scoreboard project.",
        "title": "Cricket Scoreboard API",
        "termsOfService": "https://github.com/dashu-baba/cricket-scoreboard-api",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/dashu-baba/cricket-scoreboard-api/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
        "/innings/:inningsid/over/:overid": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Innings"
                ],
                "summary": "Update an over",
                "parameters": [
                    {
                        "description": "Over Update Model",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.OverUpdateModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Innings ID",
                        "name": "inningsid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Over ID",
                        "name": "overid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
        "/innings/:inningsid/start": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Innings"
                ],
                "summary": "Created an innings",
                "parameters": [
                    {
                        "description": "Start Innings Model",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.StartInningsModel"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Innings ID",
                        "name": "inningsid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.ErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestmodels.OverUpdateModel": {
            "type": "object",
            "required": [
                "nb",
                "run"
            ],
            "properties": {
                "extra": {
                    "type": "string"
                },
                "nb": {
                    "type": "boolean"
                },
                "run": {
                    "type": "integer"
                },
                "wicket": {
                    "type": "object",
                    "$ref": "#/definitions/requestmodels.WicketDetailsModel"
                }
            }
        },
...
        "requestmodels.StartInningsModel": {
            "type": "object",
            "required": [
                "bowlerID",
                "nonStrikeBatsmanId",
                "strikeBatsmanId"
            ],
            "properties": {
                "bowlerID": {
                    "type": "string"
                },
                "nonStrikeBatsmanId": {
                    "type": "string"
                },
                "strikeBatsmanId": {
                    "type": "string"
                }
            }
        },
...
        "requestmodels.WicketDetailsModel": {
            "type": "object",
            "required": [
                "batsmanID",
                "bowlerID",
                "wicketType"
            ],
            "properties": {
                "batsmanID": {
                    "type": "string"
                },
                "bowlerID": {
                    "type": "string"
                },
                "supportID": {
                    "type": "string"
                },
                "wicketType": {
                    "type": "integer"
                }
            }
        },
        "responsemodels.ErrorModel": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}